+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 > **LoginRadius Java SDK Change Log** provides information regarding what has changed, more specifically what changes, improvements and bug fix has been made to the SDK. For more details please refer to the [LoginRadius API Documention(https://www.loginradius.com/docs/api/v2/deployment/sdk-libraries/java-library/)
 
+# Version 11.0.1
+Release on March 17, 2021
+
+## Enhancements
+
+ - Added X-Origin-IP header support. 
+ - Added 429 error code handling for "Too Many Request in a particular time frame".
+
+
 # Version 11.0.0
 Released on **Aug 21, 2020**